add CalednarPlanSeparator, refactor to use decimals instead of doubles with prices

+++ POSCore/CalendarPlanLogic/CalendarWork.cs
@@ -3,12 +3,12 @@
     public class CalendarWork
     {
         public string WorkName { get; }
-        public double TotalCost { get; }
-        public double TotalCostIncludingContructionAndInstallationWorks { get; }
+        public decimal TotalCost { get; }
+        public decimal TotalCostIncludingContructionAndInstallationWorks { get; }
         public ConstructionPeriod ConstructionPeriod { get; }
         public int EstimateChapter { get; }
 
-        public CalendarWork(string workName, double totalCost, double totalCostIncludingContructionAndInstallationWorks, ConstructionPeriod constructionPeriod, int estimateChapter)
+        public CalendarWork(string workName, decimal totalCost, decimal totalCostIncludingContructionAndInstallationWorks, ConstructionPeriod constructionPeriod, int estimateChapter)
         {
             WorkName = workName;
             TotalCost = totalCost;
