+++ POSCore/CalendarPlanLogic/Interfaces/IConstructionPeriodCreator.cs
@@ -4,6 +4,6 @@ namespace POSCore.CalendarPlanLogic.Interfaces
 {
     public interface IConstructionPeriodCreator
     {
-        ConstructionPeriod CreateConstructionPeriod(DateTime initialDate, double totalCost, double totalCostIncludingContructionAndInstallationWorks, int[] percentages);
+       ConstructionPeriod CreateConstructionPeriod(DateTime initialDate, decimal totalCost, decimal totalCostIncludingContructionAndInstallationWorks, int[] percentages);
     }
 }
